replace wget with the implementation of go
This commit works correctly.
But an error is displayed when downloading LibreSSL.
cf: https://github.com/golang/go/issues/12855
cf: https://github.com/golang/go/issues/19895
cf: https://github.com/golang/go/issues/2057
Even if an error is displayed, execution will not stop so there is no
problem.
Because we use `os.Chdir` in the program, the location where the file is
saved is the same as when `wget` is used. If you want to remove
`os.Chdir`, we needs to remove `tar`.