add CircleCI for releasing automatically
If you provided the git tag, nginx-build is released by CircleCI
automatically.
If you want to use it, you have to add the project on CircleCI
and add GITHUB_TOKEN as an environment variable.
Now, we use travis-ci. If this PR is merged, it is unnecessary.

If you are not interested in automating the release, please close it.